Fix docker development environment
Return an error in setup_dev_env.sh if the script is run as root, as the
script is meant to be used by a non-privileged user. Also add the user
created in the Dockerfile to the sudoers so that sudo can be used inside
the docker (no password needed).