Use p4lang/p4runtime bazel rules instead of external BUILD file
Bazel support was recently added to p4lang/p4runtime, so we move to
using these rules instead of a custom BUILD file.

We also move to a specific p4lang/p4runtime sha - instead of the
"master" branch - to ensure consistent builds and avoid breaking the
build because of future p4runtime changes.