LSC: Migrate users from gtl::(flat|node)_hash_(map|set) to the absl equivalents.
The Swisstable collection of hash containers is going to be released as part of Abseil (go/absl-team). In support of this, the Swisstable types have been moved into the absl:: namespace. This change is part of the effort to update google3 to correctly reference the new names.

LSC documentation: go/lsc-absl-namespace

Tested:
TGP

Tested:
    TAP found no affected targets
    http://test/OCL:215901567:BASE:215890009:1538748712765:d2dc5659
PiperOrigin-RevId: 216365200
Change-Id: I47672941a8459f6bcba428c85a36bf01ddec96fb