Add initial Tofino support to Stratum (#71)
This commit adds support for Barefoot Tofino-based switches. It adds the
following:

 * stratum/hal/lib/barefoot: a SwitchInterface implementation for
   Barefoot ASICs (BFSwitch). As of now it only supports flow
   programming, by dispatching calls to the appropriate PINode
   instance. At the moment the BFChassisManager class does not do
   anything save report port oper status events. We assume that the SDE
   has already been installed locally and we use Bazel's repository_rule
   to import the shared libraries & header files.

 * stratum/hal/bin/barefoot: builds a standalone binary (stratum_bf)
   that includes the BFSwitch and links to the Barefoot SDE
   libraries. To build this binary make sure that you export the
   BF_SDE_INSTALL environment variable and that it points to your local
   SDE install directory.

This commit also includes a few convenience scripts to test generation
of gNMI subscription messages by Stratum. Note that the moment, port
oper status events do not seem to be taken into account and the status
is always reported as "UNKNOWN".

* Change operation mode for stratum_bf to STANDALONE

* Use local repository rule for Thrift instead of symlink

* Small edit in barefoot README

* Complete support for port oper status notifciations

When performing a gNMI ONCE subscription, the switch RetrieveValue
method is called and therefore needs to be implemented. Note that gNMI
wildcard subscriptions (interfaces/interface/...) do not seem to be
supported in Stratum at the moment.