Support port configuration through PushChassisConfig for bmv2 (#129)
* Support port configuration through PushChassisConfig for bmv2

The only config parameter which is really taken into account is the
admin status, which decides whether bmv2 actually binds to the port or
not. The rest of the config is stored in a map for later retrieval.

This only implements Bmv2Switch::RetrieveValue for more request types. We now
support: OperStatus, AdminStatus, PortSpeed (dummy), NegotiatedPortSpeed
(dummy), PortCounters and AutonegStatus (dummy). More can be added later if
needed.

Fixes #111

* Fix 2 issues reported by Carmelo:

 * Generate a PORT_DOWN event when removing the port
 * Do not return garbage stats when port has been disabled