Go back to deleted MakeErrorStreamWithOutput copy constructor
MakeErrorStreamWithOutput should never be copied, because each by design
MakeErrorStreamWithOutput instance must be casted to Status one and only
once (otherwise it is a fatal error which triggers a SIGABORT). In
particular if a temporary copy is ever created by the compiler, the
error will be triggered. To use MAKE_ERROR in a lambda, one simply needs
to declare a trailing return type (as Status).

Fixes #133