Fix deadlock in Bmv2ChassisManager
There was a deadlock happening fairly often when adding a port as the
bmv2 port monitor is generating a port up event. The bmv2 port monitor
generates 4 kinds of events: added, removed, up and down. Even though we
only subscribe to up and down, as we perform an add operation, the port
monitor grabs a lock to check if there is any registered callback for
the "added" event. In the mean time, the asynchronous port monitor
thread tries to generate a port up event for a previously added port. To
generate this event, it also grabs the lock protecting the callback
list. So we end up with:
Port add thread: holds chassis lock, waits on port monitor lock
Port monitor thread: holds port monitor lock, is waiting on chassis lock
in callback

To avoid this issue altogether, we use a Channel object. Callbacks
generated by bmv2 push events to the queue, and we spawn a thread to
process them asynchronously. This has the added advantage that this
design is more similar to what is done for actual Stratum HW
implementations.

Fixes #153