Start Phal integration for Barefoot chassis manager
We now register for transceiver events with the Phal instance. Currently
we do not do anything specifial with these events except sanity check
them and call GetFrontPanelInfo (that's in order to keep this PR small,
will be included in a future PR).

We also maintain a map from node ID to another map from port ID to slot
/ port / channel. This data structure is updated by PushChassisConfig.

Finally, one notable change. I have observed that because of a
well-known Bazel issue with cc_import and because of the fact that the
default ONLP tarball we download does not include SONAME versions of the
ONLP libraries, it is not possible to run a binary linked to the ONLP
libraries. I have implemented the necessary workaround in onlp.bzl (only
for the default case, I didn't change anything for the case where the
ONLP_INSTALL env variable is set).