Forward PI DeviceMgr log messages to Stratum glog logger for stratum_bf (#311)
* Forward PI DeviceMgr log messages to Stratum glog logger for stratum_bf

Register a customer logger with DeviceMgr that sends messages to the
Stratum logger. We only do it for stratum_bf at the moment; stratum_bmv2
actually logs DeviceMgr messages with the bmv2 logging sytem, but we may
want to uniformize the 2 in the future.

We use glog "internal" LogMessage class instead of the LOG macro to be
able to control the file name and the line number displayed in the
logs. Unfortunately, DeviceMgr does not preserve that information at the
moment, but this is an improvement we can do in the future. For now, we
set the file name to "PI-device_mgr.cpp" and the line number to 0. Note
that other parts of Stratum already use LogMessage.

* Add information to the Barefoot README about turning on FEC

* Add information about p4runtime-shell to the Barefoot README