Bump version of PI, bmv2, and P4Runtime (#67)
Includes fixes to multicast group handling in PI and support for P4 logging
primitives in bmv2.

To make the PI build pass, the version of P4Runtime has been bumped to
1.1.0-rc.1, which should not break backwards-compatibility at the P4Runtime
message level.

Changes have been tested end-to-end with stratum_bmv2 on Mininet with ONOS and
Trellis, using the instructions under tools/mininet/examples/trellis. The ONOS
version used by the example was bumped to 2.2.0 which is an LTS and should be
preferred to 2.1.0.