Bump BMv2, PI, and P4Runtime to fix action profile crash in stratum_bmv2 (#244)
A bug in BMv2 caused the stratum_bmv2 process to crash when packets were
processed by an action selector left in inconsistent state. The issue is
described here:
https://github.com/p4lang/behavioral-model/issues/893

This was causing PTF tests for fabric.p4 to fail:
https://github.com/opennetworkinglab/fabric-p4test/issues/72

To fix stratum_bmv2, this change bumps the bmv2 commit, which requires bumping
PI and P4Runtime as  well. P4Runtime is updated from 1.1.0-rc1 to
backward-compatible 1.2.0-dev.