Call gRPC server Shutdown() from a separate thread
If the thread that called gRPC Server Wait() is the one that handles the
Ctrl-C, then the thread will try to re-enter the gRPC Server mutex.

This seems to work fine with pthread_mutex (behavior is undefined), but
results in deadlock in Abseil's mutex, which becomes the default in
gRPC 1.32.0.