Refactor SwitchInterface to use generic StreamMessageRequest/Response types (#536)
This PR modifies the SwitchInterface to use the generic StreamMessageRequest and StreamMessageResponse data types instead of the embedded messages like PacketIn or PacketOut. Using the encapsulating message reduces the number of nearly identical functions in the interface. Switch implementations can still decide which sub-messages to support, but extending support in one implementation does not require changes to the interface anymore.

This change also includes code changes to allow passing digests lists and acks from and to the controller for Stratum-bf, as PI does most of the work and we only connect the parts.

We do not address the topic of caching and resending digests / acks in this PR. For now this problem is pushed into the switch implementations (by passing down the acks transparently).