Auto-detect and expose BF SDE version as config_setting (#564)
This PR implements BF SDE version auto-detection and exposes the result as a config_setting so selects() can match on it. This removes the need to manually specify the --define sde_ver=x flag on the command line. Further, we can detect and reject unsupported versions.
Detection is implemented by reading the $SDE_INSTALL/share/VERSION file and then templating the value into a string_setting, a user-defined build_setting. This is the preferred way over --defines.