Add option to make Stratum-bfrt return canonical byte strings (#677)
This PR adds a new option incompatible_enable_bfrt_legacy_bytestring_responses to make Stratum-bfrt return canonical P4Runtime byte strings. For now the flag defaults to true, which means that the current behavior remains the same, that is, padded strings. In the future we will flip and eventually remove the legacy behavior.