Add `last_change` to `OperStatus` message
This PR adds the `last_change` field to the common `OperStatus` message. It is optional and contains the time stamp when the port state last transitioned. This change does not contain the actual SwitchInterface implementations for any of the platforms. We will add them in later PRs on a case-by-case basis.