Retain port state on ChassisConfig push for Stratum-bf
A `ChassisConfig` push does not necessitate discarding the port state for known ports as it does not reset the ASIC. This PR changes the behavior of the `BfChassisManager` so we retain the old state. At runtime we should no longer encounter missing ports in the `node_id_to_port_id_to_port_state_` map and can make that an error.