Make QoS queue rates in `ChassisConfig` optional on Tofino (#752)
We discovered that it can be useful to not explicitly set any maximum or minimum rates on queues. Currently, omitting the max rates will set it to zero, effectively disabling the queue. This change makes these parameters fully optional by disabling shaping/gminrate when not set. We also enhance ergonomics by making the bytes and packet rate mutually exclusive.