Create Stratum-tools package and Docker image (#771)
This PR creates the Stratum-tools Debian package and Docker image. It contains various tools and binaries related to Stratum, but not the switch runtime itself.

The new Docker container is published as `stratumproject/stratum-tools`.

- bf_pipeline_builder
- stratum_dummy
- chassis_config_migrator
- phal_cli
- ~onlp_cli~
- stratum_stub
- procmon_main
- gnmi_cli
- stratum_replay

This also deprecates the following docker images:
 - `stratumproject/stratum-replay`
 - `stratumproject/stratum-bf-pipeline-builder`
 - `stratumproject/chassis-config-migrator`
 - `stratumproject/gnmi-cli`

They will remain available for some time, but are not updated anymore.