Reject duplicate P4Runtime election IDs from multiple controllers (#857)
P4Service currently does not reject duplicate election IDs from controllers. This leads to a data corruption when a slave controller sends a MastershipUpdate with the same election ID as the current master.

This happens because we store Controllers in a set with a custom comparison function. Controllers are not hashed, but only compared by their election ID. Thus, two controllers with the same election ID are not possible.

This PR changes this by checking for insertion success. Failure indicates a duplicate and we close the stream, all according to P4RT spec.