Make file CLI flags and usage messages uniform (#894)
The CLI flags related to TLS are ambiguous whether they expect a path to a file, or the actually file content. Established convention is using `<flag>_file` for files. This PR fixes this.

It also removes the duplicate usage strings in both the docs and code, as they are auto-generated by gflags with the `--helpshort` flag. Some were already outdated and incorrect.