Remove `RETURN_ERROR` macro (#901)
This PR replaces the usage of `RETURN_ERROR()` with `return MAKE_ERROR()`. `RETURN_ERROR` seems to be part of an older style of error handling that has not been present in the original Google code drop, other recent Google projects or Abseil. In Stratum, only contributed code reintroduced the macro. The replacement is automated and should not change any logic or program behavior.
This change brings us closer to the Abseil style.