Pull SDE tars from S3 for automated post-merge Docker builds by CI (#890)
This PR pulls the SDE tars from S3 storage to enable building of BFRT related targets and pushing of docker images directly from CI. As of now we'll build and push all supported SDE versions in a post-merge to main job. If pre-merge testing all versions becomes important in the future, we can follow up on this later.