Replace P4Runtime controller management with code from PINS project (#781)
This PR replaces the Stratum-internal P4Runtime server connection management code, part of `P4Service`, with a library extracted from the PINS project. The library implements the current version of the spec with the current mastership arbitration algorithm; **Stratum will not longer pick a new master from the backup controller pool**.

Re-using existing externally maintained code should give us the benefit of future updates for cheap. This is why we strive for minimal modifications to the imported code. In particular, we do not include Stratum deps so that other can use it as well.

Not addressed here is the persistent election ID issue.


fixes #287
fixes #380