Fix workspace status script for out-of-git builds (#950)
When building from non-git source code, e.g. release tars, the `.git` directory is not present. This causes any bazel command to fail as the workspace status script exits with non-zero code. We fix this by checking for this condition and not emitting any stamping info in that case.

Fixes #949