Emit clang coverage reports for gcov version 8+ (#1026)
Updating the base OS to Debian 10 also updated the `gcov` tool. By default, `clang` emits coverage reports in the [4.8 format](https://github.com/bazelbuild/bazel/issues/5881#issuecomment-645054093), which is not compatible with the version installed now (8.3). This is fixed by adding a flag to emit reports in the right format.