Canonicalize P4RT role config PacketIn filter byte strings (#1061)
This change enforces canonicalization on incoming role config PacketIn filter strings to prevent false negative matches when comparing to canonical PacketIns later. Additionally, the filter, if set, cannot be empty.

The PacketIn filtering feature is incompatible with the `--incompatible_enable_bfrt_legacy_bytestring_responses` flag, as canonicalization is always performed in the P4Service, but only selectively in the `BfrtPacketIoManager`, hence leading to false negatives.