First stab at weekly docker build with Stratum-bfrt (#1081)
This PR explores the concept of publishing the existing weekly docker
images from CircleCI instead of jenkins.
In this first attempt, we introduce a `release_type` parameter either
set to `latest` or the date in `YY.MM.DD` form. This results in the same
docker tag structure used now.
The interval is kept short for testing, and will be fixed later if this
approach works. (Needs to be merged to master to trigger)