Expose a virtual CPU interface with Stratum-bfrt on Tofino (#1069)
This PR introduces an additional TAP interface on Tofino. Via software
processing and a special constant in PacketIn headers (`0xbf01`), it is connected to the regular
PCIe CPU interface and can be used to steer packets from the data plane
to regular apps on the host. This implements a workaround for the broken
Ethernet CPU ports on certain Edgecore Wedge switches.