Allow external Bfrt TAP interface creation and custom names (#1089)
This PR is a followup to #1069 to make it usable in more scenarios. In
particular, it is now possible to bind to an existing TAP interface
created outside of Stratum. This allows for a more persistent setup
where a Stratum restart does not remove the interface. Further, we also
check for the canonical TUN/TAP device path in addition the the Tofino
specific one. This is useful for simulated environments and tests.