Fix crash when startum restart with pipe config before controller connected.
When the controller is connected, rx_writer_ will be set through
BfrtPacketioManager::RegisterPacketReceiveWriter. However, Stratum can restart
with an existing pipe configuration and enter the 'SdeRxThreadFunc' thread to
process PacketIn before the controller is connected. In this case, rx_writer_
will be null and lead to a crash when the 'Write' callback is called.

Signed-off-by: wanghanlin <wanghanlin@corp.netease.com>